requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
camelot-py==0.9.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
distro==1.8.0
dnspython==2.4.2
python-dotenv==1.0.0
et-xmlfile==1.1.0
fastapi==0.103.2
fonttools==4.42.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
itsdangerous==2.1.2
JPype1==1.4.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib
numpy==1.26.0
opencv-python==4.8.0.76
openpyxl==3.1.2
orjson==3.9.7
packaging==23.1
pandas
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.20.0
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
tabula-py==2.8.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.5
uvicorn==0.23.2
websockets==11.0.3
